ACE.Types.Syntax

Description

Types for the syntax tree.

Documentation

data Specification

Specifications consist of a sentence coordination followed by a period and optionally one ore more subsequent specifications.

Constructors

Specification !SentenceCoord !(Maybe Specification)

data SentenceCoord

Sentences can be coordinated by and and or. And refers to the logical conjunction, while or de-notes the logical disjunction. The logical conjunction has a higher precedence than the disjunction.

Both connectors are right-associative. The expression

A or B and C or D

is therefore ordered like

A ∨ ((B ∧ C) ∨ D)

To enable more combinations, we have introduced comma-and and comma-or. These expressions reverse the order of precedence. To achieve the order

A ∨ (B ∧ (C ∨ D))

we can write

A, or B, and C or D

A sentence coordination in general consists of a sentence coordination of a lower level (thus ensuring right-associativity) optionally followed by the respective connector and a sentence coordination of the same level.

Constructors

SentenceCoord !SentenceCoord_1 !(Maybe SentenceCoord)

data SentenceCoord_1

Constructors

SentenceCoord_1 !SentenceCoord_2 !(Maybe SentenceCoord_1)

data SentenceCoord_2

Constructors

SentenceCoord_2 !SentenceCoord_3 !(Maybe SentenceCoord_2)

data SentenceCoord_3

Constructors

SentenceCoord_3 !TopicalizedSentence !(Maybe SentenceCoord_3)

data TopicalizedSentence

A topicalized sentence can start with an existential topic or a universal topic. It needs, however, not be topicalized at all but can just be an ordinary composite sentence.

Constructors

TopicalizedSentenceExistential !ExistentialTopic !(Maybe SentenceCoord)
Example: "There is a card such that the code of the card is valid."

TopicalizedSentenceUniversal !UniversalTopic !SentenceCoord
Example: "For every code there is a card such that the code belongs to it."

TopicalizedSentenceComposite !CompositeSentence
Example: "Homer is a man."

data UniversalTopic

Constructors

UniversalTopic !UniversalGlobalQuantor !N'

data CompositeSentence

Constructors

CompositeSentenceCond !ConditionalSentence

CompositeSentenceNeg !NegatedSentence

CompositeSentence !Sentence

data ConditionalSentence

Constructors

ConditionalSentence !SentenceCoord !SentenceCoord

data NegatedSentence

Constructors

NegatedSentence !SentenceCoord

data Sentence

Constructors

Sentence !NPCoord !VPCoord

data ExistentialTopic

Constructors

ExistentialTopic !ExistentialGlobalQuantor !NPCoord

data NPCoord

Constructors

NPCoordDistributed !DistributiveMarker !UnmarkedNPCoord

NPCoordUnmarked !UnmarkedNPCoord

data UnmarkedNPCoord

Constructors

UnmarkedNPCoord !NP !(Maybe UnmarkedNPCoord)

data N'

Modified noun.

Constructors

N' !(Maybe AdjectiveCoord) !N !(Maybe ApposCoord) !(Maybe NPCoord) !(Maybe RelativeClauseCoord)

data NP

Noun-phrase.

Constructors

NP !Specifier !N'

NPPro !Pronoun

NPProper !ProperName

NPVar !Variable

data N

A noun.

Constructors

N !Text

data PP

Constructors

PP !Preposition !NPCoord

data Preposition

Constructors

Preposition !Text

data ApposCoord

Constructors

ApposCoord !Apposition !(Maybe ApposCoord)

data Apposition

Constructors

AppositionVar !Variable

AppositionQuote !Quotation

data Quotation

Constructors

Quotation !Text

data Variable

Constructors

Variable !Text

data RelativeClauseCoord

Constructors

RelativeClauseCoord !RelativeClause !(Maybe (Coord, RelativeClauseCoord))

data PossessiveNPCoord

Constructors

PossessiveNPCoordGen !GenitiveNPCoord

PossessiveNPCoordPronoun !PossessivePronounCoord

data GenitiveNPCoord

Constructors

GenitiveNPCoord !GenitiveSpecifier !GenitiveN' !GenitiveTail

GenitiveNPCoordName !ProperName !GenitiveTail

data ProperName

Constructors

ProperName !Text

data PossessivePronounCoord

Constructors

PossessivePronounCoord !PossessivePronoun !(Maybe PossessivePronounCoord)

data GenitiveTail

Constructors

GenitiveTailSaxonTail !SaxonGenitiveTail

GenitiveTailCoordtail !GenitiveCoordTail

data GenitiveCoordTail

Constructors

GenitiveCoordTail !GenitiveNPCoord

data SaxonGenitiveTail

Constructors

SaxonGenitiveTail !SaxonGenitiveMarker !(Maybe (GenitiveN', SaxonGenitiveTail))

data RelativeClause

Constructors

RelativeClauseThat !VPCoord

RelativeClauseNP !NPCoord !VPCoord

RelativeClauseThatNPVP !NPCoord !VPCoord

RelativeClauseNPVP !NPCoord !NPCoord !VPCoord

RelativeClausePP !PP !NPCoord !VPCoord

data VPCoord

Constructors

VPCoord' !VP !Coord !VPCoord

VPCoordVP !VP

data GenitiveSpecifier

Constructors

GenitiveSpecifierD !Determiner

GenitiveSpecifierPPC !PossessivePronounCoord

GenitiveSpecifierN !Integer

data GenitiveN'

Constructors

GenitiveN' !(Maybe AdjectiveCoord) !N !(Maybe ApposCoord)

data VP

Constructors

VP !V'

VPNeg !Copula !V'

data V'

Constructors

V' !(Maybe AdverbCoord) !ComplV ![VModifier]

data AdverbCoord

Constructors

AdverbCoord !Adverb !(Maybe AdverbCoord)

data ComplV

Constructors

ComplVIV !IntransitiveV

ComplVPI !PhrasalIntransitiveV !PhrasalParticle

ComplVTV !TransitiveV !Compl

ComplVPV !PhrasalTransitiveV !PhrasalParticle !Compl

ComplVPV' !PhrasalTransitiveV !Compl !PhrasalParticle

ComplVDisV !DistransitiveV !Compl !Compl

ComplVPDV !PhrasalDistransitiveV !Compl !PhrasalParticle !Compl

ComplVCopula !Copula !CopulaCompl

data PhrasalTransitiveV

Constructors

PhrasalTransitiveV !Text

data PhrasalDistransitiveV

Constructors

PhrasalDistransitiveV !Text

data CopulaCompl

Constructors

CopulaComplAPC !APCoord

CopulaComplNPC !NPCoord

CopulaComplPP !PP

data APCoord

Constructors

APCoordAnd !APgrad !APCoord

APCoord !APgrad

data APgrad

Constructors

APgradAPThan !AP !NPCoord

APgradAP !AP

data AP

Constructors

APIntrans !IntransitiveAdjective

APTrans !TransitiveAdjective !PP

data TransitiveAdjective

Constructors

TransitiveAdjective !Text

data Compl

Constructors

ComplNP !NPCoord

ComplPP !PP

data PhrasalIntransitiveV

Constructors

PhrasalIntransitiveV !Text

data PhrasalParticle

Constructors

PhrasalParticle !Text

data IntransitiveV

Constructors

IntransitiveV !Text

data TransitiveV

Constructors

TransitiveV !Text

data DistransitiveV

Constructors

DistransitiveV !Text

data IntransitiveAdjective

Constructors

IntransitiveAdjective !Text

data VModifier

Constructors

VModifierVC !AdverbCoord

VModifierPP !PP

VModifierAVPP !AdverbialPP

data AdverbialPP

Constructors

AdverbialPP !Preposition !AdverbCoord

data Adverb

Constructors

Adverb !Text

data Specifier

Constructors

SpecifyDeterminer !Determiner

SpecifyPossessive !PossessiveNPCoord

SpecifyNumberP !NumberP

data AdjectiveCoord

Constructors

AdjectiveCoord !IntransitiveAdjective !(Maybe AdjectiveCoord)

data NumberP

Constructors

NumberP !(Maybe GeneralizedQuantor) !Integer

data ExistentialGlobalQuantor

Constructors

ExistentialGlobalQuantor !Copula

data ExistentialGlobalQuestionQuantor

Constructors

ExistentialGlobalQuestionQuantor !Copula

data Aux

Constructors

Do
"do"

Does
"does"

data Coord

Constructors

And
"and"

Or
"or"

data Copula

Constructors

Is
"is"

Are
"are"

data Determiner

Constructors

The
"the"

A
"a"

An
"an"

Some
"some"

No
"no"

EveryEach
"every" / "each"

All
"all"

NotEvery
"not every"

NotEach
"not each"

NotAll
"not all"

Which
"which"

data DistributiveGlobalQuantor

Constructors

ForEachOf
"for each of"

data DistributiveMarker

Constructors

EachOf
"each of"

data GeneralizedQuantor

Constructors

AtMost
"at most"

AtLeast
"at least"

MoreThan
"more than"

LessThan
"less than"

NotMoreThan
"not more than"

NotLessThan
"not less than"

data PossessivePronoun

Constructors

His
"his"

Her
"her"

HisHer
"his/her"

Its
"its"

Their
"their"

HisHerOwn
"his own" "her own" "his/her own"

ItsOwn
"its own"

TheirOwn
"their own"

Whose
"whose"

data Pronoun

Constructors

It

He

She

HeShe

Him

HerP

HimHer

They

Them

Itself

Himself

Herself

HimselfHerself

Themselves

Someone

Somebody

Something

NoOne

Nobody

NoThing

Everyone

Everybody

Everything

NotEveryone

NotEverybody

NotEverything

What

Who

Whom

WhichP

data SaxonGenitiveMarker

The Saxon genitive used for possession.

Constructors

Apostrophe
"'"

ApostropheS
"'s"

data UniversalGlobalQuantor

Constructors

ForEvery
"for every"

ForEach
"for each"

ForAll
"for all"